TdsParserStateObject: Removed unnecessary casts.

+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/TdsParserStateObject.cs
@@ -916,7 +916,7 @@ namespace Microsoft.Data.SqlClient
                 }
                 else
                 {
-                    _longlenleft = (ulong)chunklen;
+                    _longlenleft = chunklen;
                 }
             }
 
@@ -992,7 +992,7 @@ namespace Microsoft.Data.SqlClient
                 if ((ulong)(buff?.Length ?? 0) != _longlen)
                 {
                     // if the buffer is null or the wrong length create one to use
-                    buff = new byte[(int)Math.Min((int)_longlen, len)];
+                    buff = new byte[(Math.Min((int)_longlen, len))];
                 }
             }
 
@@ -1086,12 +1086,12 @@ namespace Microsoft.Data.SqlClient
 
             while (num > 0)
             {
-                int cbSkip = (int)Math.Min((long)int.MaxValue, num);
+                int cbSkip = (int)Math.Min(int.MaxValue, num);
                 if (!TryReadByteArray(Span<byte>.Empty, cbSkip))
                 {
                     return false;
                 }
-                num -= (long)cbSkip;
+                num -= cbSkip;
             }
 
             return true;
@@ -1367,7 +1367,7 @@ namespace Microsoft.Data.SqlClient
                 {
                     if (!_attentionSent)
                     {
-                        AddError(new SqlError(TdsEnums.TIMEOUT_EXPIRED, (byte)0x00, TdsEnums.MIN_ERROR_CLASS, _parser.Server, _parser.Connection.TimeoutErrorInternal.GetErrorMessage(), "", 0, TdsEnums.SNI_WAIT_TIMEOUT));
+                        AddError(new SqlError(TdsEnums.TIMEOUT_EXPIRED, 0x00, TdsEnums.MIN_ERROR_CLASS, _parser.Server, _parser.Connection.TimeoutErrorInternal.GetErrorMessage(), "", 0, TdsEnums.SNI_WAIT_TIMEOUT));
 
                         // Grab a reference to the _networkPacketTaskSource in case it becomes null while we are trying to use it
                         TaskCompletionSource<object> source = _networkPacketTaskSource;
@@ -1704,7 +1704,7 @@ namespace Microsoft.Data.SqlClient
                 {
                     stateObj.SetTimeoutStateStopped();
                     Debug.Assert(_parser.Connection != null, "SqlConnectionInternalTds handler can not be null at this point.");
-                    AddError(new SqlError(TdsEnums.TIMEOUT_EXPIRED, (byte)0x00, TdsEnums.MIN_ERROR_CLASS, _parser.Server, _parser.Connection.TimeoutErrorInternal.GetErrorMessage(), "", 0, TdsEnums.SNI_WAIT_TIMEOUT));
+                    AddError(new SqlError(TdsEnums.TIMEOUT_EXPIRED, 0x00, TdsEnums.MIN_ERROR_CLASS, _parser.Server, _parser.Connection.TimeoutErrorInternal.GetErrorMessage(), "", 0, TdsEnums.SNI_WAIT_TIMEOUT));
 
                     if (!stateObj._attentionSent)
                     {
@@ -2680,7 +2680,7 @@ namespace Microsoft.Data.SqlClient
                 try
                 {
                     Debug.Assert(_parser.Connection != null, "SqlConnectionInternalTds handler can not be null at this point.");
-                    AddError(new SqlError(TdsEnums.TIMEOUT_EXPIRED, (byte)0x00, TdsEnums.MIN_ERROR_CLASS, _parser.Server, _parser.Connection.TimeoutErrorInternal.GetErrorMessage(), "", 0, TdsEnums.SNI_WAIT_TIMEOUT));
+                    AddError(new SqlError(TdsEnums.TIMEOUT_EXPIRED, 0x00, TdsEnums.MIN_ERROR_CLASS, _parser.Server, _parser.Connection.TimeoutErrorInternal.GetErrorMessage(), "", 0, TdsEnums.SNI_WAIT_TIMEOUT));
                     _bulkCopyWriteTimeout = true;
                     SendAttention();
                     _parser.ProcessPendingAck(this);
